==> 0.1_Perimeter Cleaner.ipynb <==
Where there are two or more perimeters for the same day, the perimeter cleaner combines them using a logical OR.

import numpy as np
import os
import pickle
from collections import defaultdict
import pyprind

import matplotlib.pyplot as plt

def filename_map(files):
    fn_map = defaultdict(list)
    for file in files:
        components = file.split(' ')
        fn_map[components[0]+' '+components[1]].append(file)
        
    return(fn_map)

perims =  os.listdir('perims/dirty')

fn_map = filename_map(perims)

errors = []

progbar = pyprind.ProgBar(len(fn_map))

for firedate in fn_map:
    files = fn_map[firedate]
    try:
        op = open('perims/dirty/'+files[0], 'rb')
        f = pickle.load(op)
        r = f['raster']
        op.close()

        for file in files:
            op = open('perims/dirty/'+file, 'rb')
            r2 = pickle.load(op)['raster']
            op.close()
            r = np.logical_or(r, r2, dtype=np.int8)

        f['raster'] = r

        op = open('perims/clean/'+firedate+'.pkl', 'wb')
        pickle.dump(f, op)
        op.close()
    except:
        errors.append(files)
        pass
    progbar.update()
