==> 1_ArcGIS code.ipynb <==
ArcGIS has a built in Python interpreter (the only saving grace of that piece of software).  This code is copy/pasted into ArcGIS to extract rasters for topology and fuel load matching the fire perimeters.

import pickle
import numpy as np
import pandas as pd

df = pickle.load(open('2012_fire_bounds.pkl', 'rb'))
df.head()

import pandas as pd

layers = ['us_slp', 'us_dem', 'us_asp', 'us_130fbfm40']
#'us_130fbfm40'

def get_neighborhood(layer, left, bottom, ncols, nrows):

    offset= 0
    left = left +offset
    bottom = bottom+offset

    a = arcpy.RasterToNumPyArray(
        layer,
        arcpy.Point(left, bottom),
        ncols,
        nrows
    )
    return a

def create_array(layers, name, left, bottom, ncols, nrows):
    data = dict()
    data['fire_id'] = name
    data['ref_point'] = (left, bottom)
    data['shape'] = (ncols, nrows)
    data['type'] = 'landfire'
    
    for layer in layers:
        data[layer]=get_neighborhood(layer, left, bottom, ncols, nrows)
    
    fn = name+ ' topo.pkl'
    outfile = open('topo/'+fn,'wb')
    pickle.dump(data,outfile)
    outfile.close()

df.sort_values('a', inplace=True)

for row in df.itertuples():
    name = row[1]
    left =row[2]
    bottom=row[5]
    ncols = row[6]
    nrows = row[7]
    
    print(name, row[8])
    create_array(layers, name, left, bottom, ncols, nrows)
