==> 2_Wind Getter.ipynb <==
NREL WIND is a federal database containing extremely detailed weather data for 2007-2012 accessible through a HSDS API.
https://developer.nrel.gov/docs/wind/wind-toolkit/

%matplotlib inline
import h5pyd
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import matplotlib.image as mpimg
import pyproj
from pyproj import Proj
import dateutil
import pymongo as pm
import pyprind
import gridfs
import json
import pickle
import geopandas as gpd
import os

h5pyd.getServerInfo()

h5pyd.Config()

f = h5pyd.File("/nrel/wtk-us.h5", 'r')

#load our datasets

windspeed = f['windspeed_10m']
winddirection = f['winddirection_10m']
precipitation = f['precipitationrate_0m'] 
relative_humidity = f['relativehumidity_2m']
temp = f['temperature_2m']
dt = f["datetime"]
dt = pd.DataFrame({"datetime": dt[:]},index=range(0,dt.shape[0]))
dt['datetime'] = dt['datetime'].apply(dateutil.parser.parse)

dt.tail()

twentytwelve = dt.loc[(dt.datetime >= '2012-01-01') & (dt.datetime < '2013-01-01')].index
twentytwelve

timestep = dt.loc[dt.datetime == '2012-04-01 12:00:00'].index[0]
timestep

NREL functions for going from lat-lon to WIND internal coordinates, which are 1 km squares.

def indicesForCoord(f, lat_index, lon_index):
    dset_coords = f['coordinates']
    projstring = """+proj=lcc +lat_1=30 +lat_2=60 
                    +lat_0=38.47240422490422 +lon_0=-96.0 
                    +x_0=0 +y_0=0 +ellps=sphere 
                    +units=m +no_defs """
    projectLcc = Proj(projstring)
    origin_ll = reversed(dset_coords[0][0])  # Grab origin directly from database
    origin = projectLcc(*origin_ll)
    
    coords = (lon_index,lat_index)
    coords = projectLcc(*coords)
    delta = np.subtract(coords, origin)
    ij = [int(round(x/2000)) for x in delta]
    return tuple(reversed(ij))

def indicesForBBox(f, lat_range, lon_range):
    xmin = None
    xmax = None
    ymin = None
    ymax = None
    for i in [0,1]:
        for j in [0,1]:
            yx = indicesForCoord(f,lat_range[i],lon_range[j])
            if xmin is None or yx[1] < xmin:
                xmin = yx[1]
            if xmax is None or yx[1] > xmax:
                xmax = yx[1]
            if ymin is None or yx[0] < ymin:
                ymin = yx[0]
            if ymax is None or yx[0] > ymax:
                ymax = yx[0]
            if xmin == xmax:
                xmax += 1
            if ymin == ymax:
                ymax +=1
    return ([xmin,xmax],[ymin,ymax])

df = pickle.load(open('2012_fire_bounds.pkl', 'rb'))
df.head()

epsg5070 = pyproj.Proj("+init=epsg:5070")
epsg4269 = pyproj.Proj("+init=epsg:4269")

NREL WIND only allows 500 requests an hour, so this takes two sets of requests to run.  We iterate through the unique fires database, check against the `done` set, and then dump to a pickle file.

unique_fires =  set(df.fire_id.unique())
print(len(unique_fires))

done =  set(os.listdir('wind'))
    
unique_fires = unique_fires-done

print(len(unique_fires))



a= [x for x in np.arange(len(df)-1, -1, -1)]

for k in a:
    fire = df.iloc[k] #this is not the preferred way to access a df, but it works
    
    fire_id = fire.fire_id
    
    done = os.listdir('wind')
    if fire_id +' wind.pkl' not in done:

        lon1=df.iloc[k,1]
        lat1=df.iloc[k,3]

        lon2 = df.iloc[k,2]
        lat2 = df.iloc[k,4]

        xs = []
        ys = []

        x,y = pyproj.transform(epsg5070, epsg4269, lon1, lat1)
        xs.append(x)
        ys.append(y)
        x,y = pyproj.transform(epsg5070, epsg4269, lon1, lat2)
        xs.append(x)
        ys.append(y)
        x,y = pyproj.transform(epsg5070, epsg4269, lon2, lat1)
        xs.append(x)
        ys.append(y)
        x,y = pyproj.transform(epsg5070, epsg4269, lon2, lat2)
        xs.append(x)
        ys.append(y)

        bounds = [min(ys), max(ys), min(xs), max(xs)]

        x_range, y_range =indicesForBBox(f, (bounds[0], bounds[1]), (bounds[2], bounds[3]))

        fire_dates =  df2[df2.uniquefire==fire_id]    
        enddate =  fire_dates.iloc[0].date
        startdate = fire_dates.iloc[-1].date

        (startdate, enddate)

        t0 = dt.loc[dt.datetime == startdate].index[0]-100
        tf = dt.loc[dt.datetime == enddate].index[0]

        local_ws = windspeed[t0:tf, y_range[0]:y_range[1], x_range[0]: x_range[1]]

        local_wd = winddirection[t0:tf, y_range[0]:y_range[1], x_range[0]: x_range[1]]

        local_precip = precipitation[t0:tf, y_range[0]:y_range[1], x_range[0]: x_range[1]]

        local_humidity = relative_humidity[t0:tf, y_range[0]:y_range[1], x_range[0]: x_range[1]]

        local_temp = temp[t0:tf, y_range[0]:y_range[1], x_range[0]: x_range[1]]

        shape = local_ws.shape

        data = {
                'fire_id':fire_id,
                'startdate': startdate,
                'enddate': enddate,
                'shape':  shape,
                'bounds' : bounds,
                'windspeed' : local_ws,
                'winddirection' : local_wd,
                'precipitation': local_precip,
                'humidity' : local_humidity,
                'temperature' : local_temp
                }

        outfile = open('wind/'+ fire_id +' wind.pkl','wb')
        pickle.dump(data,outfile)
        outfile.close()
